+++ changelog.md
@@ -8,6 +8,7 @@
     - Keyboard support, Use numpad or arrow keys to scroll, delete key to remove a room, + or - to change level, / or * to change zone
     - Mouse scroll lock, double clicking will lock scroll mode until click or escape ends it.
   - Code editor:
+    - Add hover help for efuns, sefuns, and other supported functions that have supplied help data
     - Add basic Go to/Peek definition, works only with #define macros and only if in current file or included files relative to file
     - Format set_aggressive better when generating monster code
     - Virtual area editor: